requirements.txt
appnope==0.1.0
blessings==1.6
bpython==0.16
certifi==2017.7.27.1
chardet==3.0.4
curtsies==0.2.11
decorator==4.1.2
Django==1.11.4
django-annoying==0.10.3
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-location-field==2.0.3
django-map-widgets==0.1.9
django-npm==1.0.0
djangorestframework==3.6.4
-e git+https://github.com/mobolic/facebook-sdk.git@410f5c5892acea0c7a45a5382abccb6b7c702d6d#egg=facebook_sdk
geopy==1.11.0
googlemaps==2.5.1
greenlet==0.4.12
icalendar==3.11.7
idna==2.6
ipython-genutils==0.2.0
jedi==0.10.2
nodeenv==1.2.0
pexpect==4.2.1
pickleshare==0.7.4
pprint==0.1
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.22
wcwidth==0.1.7
